+++ app/Http/Controllers/FuentesPrimarias/TipoRespuestaController.php
@@ -9,6 +9,11 @@ use App\Http\Requests\ModificarTipoRespuestaRequest;
 use App\Models\Estado;
 use App\Models\PonderacionRespuesta;
 use App\Models\TipoRespuesta;
+use App\Models\BancoEncuestas;
+use App\Models\Encuesta;
+use App\Models\Proceso;
+use App\Models\PreguntaEncuesta;
+use App\Models\Pregunta;
 use DataTables;
 class TipoRespuestaController extends Controller
 {
@@ -147,11 +152,43 @@ class TipoRespuestaController extends Controller
     public function destroy($id)
     {
         $tipoRespuestas = TipoRespuesta::findOrFail($id);
-        $tipoRespuestas->delete();
-        return response([
-            'msg' => 'El tipo de respuesta ha sido eliminado exitosamente.',
-            'title' => '¡Tipo de respuesta Eliminado!'
-        ], 200)// 200 Status Code: Standard response for successful HTTP request
-        ->header('Content-Type', 'application/json');
+        $preguntas = Pregunta::where('FK_PGT_TipoRespuesta','=',$tipoRespuestas->PK_TRP_Id)->get();
+        $contador = 0;
+        foreach($preguntas as $pregunta)
+        {
+            $preguntas_encuestas = PreguntaEncuesta::where('FK_PEN_Pregunta','=',$pregunta->PK_PGT_Id)->get();
+            foreach($preguntas_encuestas as $pregunta_encuesta)
+            {
+                $banco_encuestas = BancoEncuestas::findOrFail($pregunta_encuesta->FK_PEN_Banco_Encuestas);
+                $encuestas = Encuesta::where('FK_ECT_Banco_Encuestas','=',$banco_encuestas->PK_BEC_Id)
+                ->get();
+                foreach($encuestas as $encuesta)
+                {
+                    $proceso = Proceso::find($encuesta->FK_ECT_Proceso);
+                    if ($proceso->FK_PCS_Fase == 4)
+                        $contador = 1;
+                        break;     
+                }
+                if($contador ==1) break;
+            }
+            if($contador==1) break;
+        }
+        if($contador==0)
+        {
+            $tipoRespuestas->delete();
+            return response([
+                'msg' => 'El tipo de respuesta ha sido eliminado exitosamente.',
+                'title' => '¡Tipo de respuesta Eliminado!'
+            ], 200)// 200 Status Code: Standard response for successful HTTP request
+            ->header('Content-Type', 'application/json');
+        }
+        else
+        {
+            return response([
+                'errors' => ['Existe una pregunta que hace parte de una encuesta en uso para un proceso que se encuentra en fase de captura de datos'],
+                'title' => '¡Error!'
+            ], 422)// 200 Status Code: Standard response for successful HTTP request
+                ->header('Content-Type', 'application/json');
+        }
     }
 }
